Tp 3 JJWT, se permite acceso a ciertos endpoints solo para usuarios logueados, los duenos de las playlist pueden modificarlas unicamente.

+++ src/main/java/ar/edu/unnoba/pdyc/mymusic/service/PlaylistServiceImp.java
@@ -1,13 +1,16 @@
 package ar.edu.unnoba.pdyc.mymusic.service;
 
 import ar.edu.unnoba.pdyc.mymusic.model.Song;
+import ar.edu.unnoba.pdyc.mymusic.model.User;
 import ar.edu.unnoba.pdyc.mymusic.repository.PlaylistRepository;
 import ar.edu.unnoba.pdyc.mymusic.model.Playlist;
+import ar.edu.unnoba.pdyc.mymusic.repository.UserRepository;
 import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Objects;
 
 @Service
 public class PlaylistServiceImp implements PlaylistService {
@@ -18,40 +21,57 @@ public class PlaylistServiceImp implements PlaylistService {
     @Autowired
     private SongService songService;
 
+    @Autowired
+    private UserRepository userRepository;
+
     @Override
-    public void createPlaylist(String name) {
-        Playlist playlist = new Playlist(name);
+    public void createPlaylist(String name, String mail) {
+        User owner = userRepository.findByEmail(mail);
+        Playlist playlist = new Playlist(name, owner);
         repository.save(playlist);
     }
 
     @Override
-    public void deletePlaylist(Long playlistId) {
+    public void deletePlaylist(Long playlistId, String mail) {
+        Playlist playlist = repository.findById(playlistId).get();
+        if (!isOwner(playlist, mail))
+            throw new RuntimeException("You are not the owner of the playlist");
         repository.deleteById(playlistId);
     }
 
-    @Transactional  // This annotation is needed to avoid LazyInitializationException
+    @Transactional
     @Override
-    public void addSongToPlaylist(Long songId, Long playlistId) {
+    public void addSongToPlaylist(Long songId, Long playlistId, String mail) {
         Song song = songService.getSong(songId);
         Playlist playlist = repository.findById(playlistId).get();
-        playlist.getSongs().size(); // Initialize the songs collection
+        if(!isOwner(playlist, mail))
+            throw new RuntimeException("You are not the owner of the playlist");
+        if (validateSongPresence(playlist, song))
+            throw new RuntimeException("The song is already in the playlist");
+        playlist.getSongs().size();
         playlist.addSongToPlaylist(song);
         repository.save(playlist);
     }
 
     @Transactional
     @Override
-    public void removeSongFromPlaylist(Long songId, Long playlistId) {
+    public void removeSongFromPlaylist(Long songId, Long playlistId, String mail) {
         Playlist playlist = repository.findById(playlistId).get();
         Song song = songService.getSong(songId);
+        if(!isOwner(playlist, mail))
+            throw new RuntimeException("You are not the owner of the playlist");
+        if (!validateSongPresence(playlist, song))
+            throw new RuntimeException("The song is not in the playlist");
         playlist.getSongs().remove(song);
         repository.save(playlist);
     }
 
     @Override
-    public void updatePlaylistName(Long playlistId, String name) {
+    public void updatePlaylistName(Long playlistId, String nameSong, String nameOwner) {
         Playlist playlist = repository.findById(playlistId).get();
-        playlist.setName(name);
+        if (!isOwner(playlist, nameOwner))
+            throw new RuntimeException("You are not the owner of the playlist");
+        playlist.setName(nameSong);
         repository.save(playlist);
     }
 
@@ -60,17 +80,16 @@ public class PlaylistServiceImp implements PlaylistService {
     public List<Playlist> getPlaylists() {
         List<Playlist> playlists = repository.findAll();
         for (Playlist playlist : playlists) {
-            playlist.getSongs().size(); // Forzamos la carga de las canciones para evitar LazyInitializationException en el recurso PlaylistResource cuando se mapea a PlaylistDTO en el método mapToPlaylistInfo de PlaylistResource
+            playlist.getSongs().size();
         }
         return playlists;
     }
-
 
     @Transactional
     @Override
     public List<Song> getSongsFromPlaylist(Long playlistId) {
         Playlist playlist = repository.findById(playlistId).get();
-        playlist.getSongs().size(); // Forzamos la carga de las canciones
+        playlist.getSongs().size();
         return playlist.getSongs();
     }
 
@@ -78,4 +97,16 @@ public class PlaylistServiceImp implements PlaylistService {
     public Playlist getPlaylistByID(Long playlistId) {
         return repository.findById(playlistId).get();
     }
+
+    private boolean isOwner(Playlist playlist, String mail) {       //retorna true si el mail pasado por parametro es igual al mail del propietario de la playlist
+        return Objects.equals(playlist.getOwner().getEmail(), mail);
+    }
+
+    private boolean validateSongPresence(Playlist playlist, Song song) {    //retorna true si la playlist contiene la cancion
+        return playlist.getSongs().contains(song);
+    }
+
+
 }
+
+
